:recycle: chore(api) extract setup configurations seeds
