+++ api/db/seeds/data/team-certification/data-builder.js
@@ -10,39 +10,23 @@ import {
   V3_CERTIFICATION_CENTER_USER_ID,
   V3_PRO_PILOT_EXTERNAL_ID,
 } from './constants.js';
-import { createCompetenceScoringConfiguration } from './create-competence-scoring-configuration.js';
-import { createIssueReportCategories } from './create-issue-report-categories.js';
 import { proOrganizationWithCertifCenter } from './create-pro-organization-with-certif-center.js';
 import { scoOrganizationManaginAgriStudentsWithFregata } from './create-sco-organization-managing-agri-student-with-fregata.js';
 import { scoOrganizationManaginStudentsWithSiecle } from './create-sco-organization-managing-student-with-siecle.js';
 import { scoOrganizationNotManagingStudents } from './create-sco-organization-not-managing-students.js';
-import { createScoringConfiguration } from './create-scoring-configuration.js';
+import { setupConfigurations } from './setup-configuration.js';
 
 async function teamCertificationDataBuilder({ databaseBuilder }) {
   await scoOrganizationManaginAgriStudentsWithFregata({ databaseBuilder });
   await scoOrganizationManaginStudentsWithSiecle({ databaseBuilder });
   await proOrganizationWithCertifCenter({ databaseBuilder });
   await scoOrganizationNotManagingStudents({ databaseBuilder });
-  _createV3CertificationConfiguration({ databaseBuilder });
-  createCompetenceScoringConfiguration({ databaseBuilder });
-  createScoringConfiguration({ databaseBuilder });
   await _createV3PilotCertificationCenter({ databaseBuilder });
   await _createSuccessCertifiableUser({ databaseBuilder });
-  await createIssueReportCategories({ databaseBuilder });
+  await setupConfigurations({ databaseBuilder });
 }
 
 export { teamCertificationDataBuilder };
-
-function _createV3CertificationConfiguration({ databaseBuilder }) {
-  databaseBuilder.factory.buildFlashAlgorithmConfiguration({
-    maximumAssessmentLength: 32,
-    challengesBetweenSameCompetence: null,
-    limitToOneQuestionPerTube: true,
-    enablePassageByAllCompetences: true,
-    variationPercent: 0.5,
-    createdAt: new Date('1977-10-19'),
-  });
-}
 
 async function _createV3PilotCertificationCenter({ databaseBuilder }) {
   databaseBuilder.factory.buildUser.withRawPassword({
